Add controllers, routes, serializers, and tests for them

+++ db/migrate/20210716164348_create_flights.rb
@@ -3,7 +3,7 @@ class CreateFlights < ActiveRecord::Migration[6.1]
     create_table :flights do |t|
       t.string :name, null: false
       t.integer :no_of_seats
-      t.float :base_price, null: false
+      t.integer :base_price, null: false
       t.datetime :departs_at, null: false
       t.datetime :arrives_at, null: false
       t.belongs_to :company, index: true, foreign_key: true
