+++ db/migrate/20210716164849_create_bookings.rb
@@ -2,7 +2,7 @@ class CreateBookings < ActiveRecord::Migration[6.1]
   def change
     create_table :bookings do |t|
       t.integer :no_of_seats, null: false
-      t.float :seat_price, null: false
+      t.integer :seat_price, null: false
       t.belongs_to :user, index: true, foreign_key: true
       t.belongs_to :flight, index: true, foreign_key: true
       t.timestamps
